Import SpinLock from gperftools
This is the spin lock implementation used in tcmalloc/gperftools, which
should be better than the boost simple_spinlock for the following reasons:
- adaptively falls back to futex after spinning for a while
- supports lock contention profiling (the code for this isn't open, but
  the hook is there and we can plug into it later)
- seems to do a smarter thing about keeping the fast path inline and the
  slow path out-of-line, so that the generated code is a little less bulky
  at the call sites

This imports some other utility code (eg sysinfo.cc). I pared this file
down to only include the required functions. Additionally, had to wrap
things in a base:: namespace to avoid conflicting with gperftools, which
is sloppy about managing its exported symbols.

This also imports valgrind.h, which is BSD-licensed, and necessary to
implement RunningOnValgrind().

Change-Id: I3fcd88cde1cfd246f2ee6437e2c3560e9e6d6c7d
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1220
Tested-by: jenkins
Reviewed-by: Colin McCabe <cmccabe@cloudera.com>