spinlock: automatically add trace entries when lock contention is experienced
This adds a hook to the google SpinLock contention code which, if the lock
acquisition time exceeds a specified threshold, and there is a current trace
available, it will log a trace message.

The trace message includes the lock acquisition time in nanoseconds as well as
the hex stack trace from the thread which got delayed.

This includes a basic test of the trace message. I also manually tested by
running a tablet server and hitting it with a lot of load. I saw the following
in the log:

0104 23:17:40.270636 Waited 4856768ns on lock 0x1f34d00. stack: 00000000009ab81d 00000000009ab945 00000000009df91e 00000000007ff8db 00000000007ff910 000000000080012b 00000000007fec09 00000000007fd8c0 000000000078d89f 0000000000a9ca6a 00000000008ea21e 00000000008eb209 00000000008eb178 00000000008eb125 00000000008eb0e4 00007f4347900d47 0000003c98607851 0000003c982e894d

Pasting the hex addresses into "addr2line -e build/latest/kudu_tablet-server"
resulted in the nice human-readable stacktrace with line numbers, etc.

Note that this commit changes the namespace of the spinlock contention
submission function from the 'base' into 'gutil' namespace. Without this,
tcmalloc was submitting its own lock contention into our code. But, our code
may cause allocations, which aren't safe to do from within the tcmalloc path
(since it may hold an allocator-wide lock). Moving to our own namespace loses
the advantage of tracing tcmalloc's locks, but makes it so we can safely
allocate within the profiling code.

Future directions from here:
- ability to associate a lock object with a metric, to collect contention count
  and distribution of acquisition time
- import code from gperftools which collects counts grouped by stack traces,
  and set up support for pprof contention profiling via HTTP.

Change-Id: I97b9cffde7a9f87675dacb0a4c8b8836fe650e25
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1223
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>
Tested-by: jenkins