Return data on open scanner, close automatically
By default, scanners will call next after their creation.
It's possible to do the old-style open by calling SetBatchSizeBytes(0).
This patch also adds auto-close when the user sets close_scanner
on the ScanRequest and it won't read any data if SetBatchSizeBytes
was set to 0.
These changes enable single-RPC scans for short ranges of rows since
the Open call will both fetch the data and close the scanner on the TS
if the range predicate returns false.

Change-Id: Ie017e758f04da3323b0e99bcf62aa972584d307f
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1228
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>