rpc: add the ability to respond to RPCs with application-specific errors
This adds a feature to the RPC framework to make it easier to pass
application-specific error responses back to the user. Prior to this patch,
there was no RPC-system support for "error" responses, and we relied on all of
the RPC response messages having some manually-assembled 'error' code. This
made for funny looking code on the server which would set an error field and
then "RespondSuccess()", rather unintuitively. On the client side, the code
would unintuitively make an RPC, see a success Status, and then have to
remember to check response.has_error().

Now, the user of the framework can register a protobuf "extension" with
kudu.rpc.ErrorStatusPB, and then use the new utility function
RpcContext::RespondApplicationError(...) to respond with an APPLICATION_ERROR
code. On the client side, this causes a RemoteError status to be returned, and
the user can fetch the detailed status protobuf from the controller.

See the changes to the tests for example usage.

A follow-up commit will change the existing services to use this new error-passing
mechanism.

Change-Id: I83834cb9e411b09e3787a2a242ac23b5e3d1627a
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1230
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>