rpc: clean up RPC error codes and specify them more often
This removes some of the RPC response error codes which were unused,
and adds some new ones which make more sense. Additionally, the error
codes are actually now returned in the majority of situations, and
the APIs are modified to make it hard to respond to an RPC without
setting some kind of code.

Change-Id: Iadc22eee5815434a6f762c4181752507ef95f7c1
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1231
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>