gutil: replace the word 'Kudu' with 'Cloudera'
Impala has imported gutil into their repo, but hasn't yet pushed it
to public github. Henry noticed that we use the code name "Kudu"
in the source. He's making an equivalent change to replace this with
'Cloudera' in the Impala tree. Making the same change here so that
we don't carry the diff forward.

Change-Id: Ieed92a9dfc6370b80b6353e0cfc3f6ed8156abec
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1234
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>