build: enable -fno-omit-frame-pointer in RELEASE build
This has a slight hit to performance, but means that profilers that
use the frame pointer to walk the stack will actually work properly.
This includes tools like "perf" and the gperftools cpu profiler.

Given we're a new project, I think we'll get more performance mileage
in the long run from making it easy to profile than we will from
enabling this option. If we find down the line that we don't use these
tools much, we can consider reverting this to boost back performance
a percent or two.

Change-Id: I224d6ca6c616ae39e1681f3d11e79f63ed94030f
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1238
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>
Tested-by: jenkins