tablet: pass around pointers to Schemas instead of references
A lot of parts of the code were passing const Schema& around, and then
some short-lived objects were actually holding onto references as members.
Having reference-typed members is something we should try to avoid, since
it makes it non-obvious to the caller that the argument is going to be
captured and has some kind of lifetime requirements.

In a couple places we were _not_ doing the reference member, but probably
could have. For example, the DeltaProjector class and cfile iterators
made Schema copies when it would have been sufficient for them all
to hold onto a pointer to the top-level Tablet iterator's projection
object.

I anticipate this will improve performance for some workloads - particularly
those with lots of short scans.

Change-Id: Ia450909602c778b3f042c348bf796ef2e30ede12
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1244
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>