kudu-lint: add AST matcher to find class members which are reference-typed
Reference-typed class members are a bad practice (of which I am probably the
most frequent offender in the code base). They are typically instantiated
from some object passed into the constructor by a caller, and it may be
non-obvious to the caller that the passed-in object needs to stay valid.

This addition to the clang-based lint tool automatically finds any cases
where we have reference-typed class members and reports them as errors.

Change-Id: I9353df424ced2839bb633ccfcff9b116db52acba
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1246
Tested-by: jenkins
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>