partial_row: add support for projecting a partial row into a contiguous server row
This will be used in tablet_peer to decode inserts provided by the client.
This patch takes the partial row (which may have unset or missing columns)
and projects it into the in-memory contiguous format expected by Insert(),

This patch just adds the utility function, and does not yet integrate it into
the wire protocol or tablet itself.

Change-Id: I37e84283a3b5057d8482be2ce113ca7d0790e753
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1259
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>