tablet_server: run maintenance threads on all local tablets
Prior to this patch, the maintenance threads in the tablet server were
hard-coded to run against the single demo tablet that was created on the
server. When tables are added/dropped dynamically, this doesn't work, since the
tablet doesn't exist at startup.

This patch changes the threads to enumerate the tablets from the
TSTabletManager and round-robin through performing maintenance on each, or
doing nothing if there are no tablets.

Change-Id: I88e025e9a7767249b07835bd293d0696c108e7f4
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1269
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins