KUDU-89. Fix heap-use-after-free in RPC trace
This fixes a heap-use-after-free which is seen in patterns like:

  lock_guard<spinlock> l(lock_);
  ... do some stuff ...
  rpc->RespondSuccess();
  // lock goes out of scope

The issue with the above pattern is that the lock going out of scope can
produce trace entries, and the threadlocal trace is still set to point to
the RPC's trace. The trace may get freed, however, any time after the
RPC is Responded to.

The fix in this patch is to make Trace objects ref-counted so that the
RPC response doesn't immediately free it, if there are still handler
threads potentially hanging onto references in their thread locals.

Change-Id: I55bba95df2ec49fd3f7b5fd0a7c8b4a65294b2df
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1284
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>