trace: add filename and line number to trace messages
This makes it easier to figure out where a given trace message is coming from.
The format for the new info is the same as glog.

This commit also removes Trace::Message() since its functionality is a subset
of Trace::SubstituteAndAppend. All tracing call sites are also changed to use
macros so that the file/line information is available.

Change-Id: Ia3c23e1da4cf55fa6ce60ad2551eccc2b3aca92b
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1286
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>