threadpool: fix reference counting in trace propagation
When the Trace pointer was propagated into the queue entry, its reference
count wasn't incremented. This meant that, if the submitting thread finished
before the task was able to run, the task would end up picking up the
trace after its reference count had potentially dropped to 0, causing
a use-after-free violation.

This simply increments the refcount upon adding to the threadpool queue
and decrements when it is removed.

Change-Id: I6c393099c37041667f8d7f202f10d88cba21fdc6
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1320
Tested-by: David Alves <david.alves@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins