Add RWCLock, a read-write-commit lock
This will be used in the master for TabletInfo/TableInfo. These structures need
to be committed to disk when they are mutated, but we do not want to block
readers during the lengthy disk-write process. Using this type of lock, we can
do the mutation in a copy-on-write fashion, then hold the commit lock only long
enough to swap in the mutated data.

Verified the unit test coverage by commenting out a couple of the "wait" loops
in the implementation and ensuring that the test failed with the broken code.

Change-Id: I50f3ac3d2077a6b52a12695844aaba531853958a
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1324
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>