Support compression, encoding for DDL purposes
Implemented a way to specify encoding an compression type. If the
specified encoding type is not found, we log a message and fall back
to the default encoding type.

TODO: Once AlterTable is implemented, fail alter if invalid encoding
is specified.

TODO: Once we settle on a way on handling internal attributes,
implement a cleaner API for specifiying attributes

Change-Id: If87d3b2ef6887003a15fa7120c50e30f2b59b4a8
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1374
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>