client: allow client to insert partial rows
This removes the check in the client that was ensuring that all columns are
fully specified, and modifies the tests to set up a schema where there's an
extra column which we never explicitly write to. The tests now verify that this
extra column ends up filled in with its default.

Also fixed a bug where SchemaToPB was dropping the 'write default' field
if there was a read default set. This behavior didn't make sense, since
it is actually possible to have a read default without a write default
(eg if a column is added with a default, and then the default is removed,
we still need the read default in order to materialize data for the
columns in older disk rowsets)

Change-Id: Id36d47ee70bce184cfc9d12c6607a34a51ba2d39
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1377
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>