Master side Alter Table
The Master send, and keeps retrying, the alter schema
to each TS with a running tablet.
In case the TS is down or goes down, the alter schema
request is sent after receiving the heartbeat with the
wrong "schema version".

The client can poll IsAlterTableDone() to verify if an
alter table is in-progress or not.

Change-Id: I9336c7942fc22193c915e0f954917fa87f2cd4d8
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1277
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>