Change wire protocol to use RowOperationsPB for update and delete
This patch gets rid of the expectation that the client uses the internal
RowChangeList format to send updates, and instead has them send RowOperationsPB.

All of the operations are collected together in a single operation list, and
similarly all of the results are consolidated into a single result list in the
commit record.

In the process of making these changes, also changed some of the write
transaction code to pass around RowChangeLists by value instead of allocating
and passing by pointer. These objects are only a pointer and length, so copying
is likely going to be cheaper than the malloc/free pair.

This results in a substantially cleaner bootstrap as well, since we can loop
over all of the operations together.

Change-Id: I8e84398e0dd681e6e3c2183065a9b5650843907d