Refactor serialization and deserialization out of PartialRow class, clean it up
This mostly moves the code for serializing/deserializing RowOperationsPB out of
PartialRow into new Encoder/Decoder classes.

Along the way, also consoliated the code for INSERT/UPDATE/DELETE to share the
majority of code, and store the decoded operations in structs so that a single
block can return INSERT, UPDATE, and DELETE mixed together. This is the main
reason for this patch: we'd like to consolidate all of the operations in a
single PB so that the "per_row_errors" code can point to an individual
operation index without too much complexity.

Functionally, the code is mostly the same except that we now check that the
client schema has no IDs and that the server schema does have IDs. The
consolidated code also mishandles one rare case: where a client has an old
schema which is missing a new column which is required, but is trying to
perform an UPDATE which doesn't touch that column. This is not actually
a regression - the tests are just more accurately using the APIs that the
real code uses, and thus this test case fails where it used to pass.

Will file a JIRA for this issue to match the TODO.

Change-Id: I563d54fdbc4ebec2cd6a96cebe12ab26a5db3b58
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1401
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>