Fix bug where first row of MRS would not be included if MRS was part of compaction
The MergeCompactionInput code was calling FinishBlock on its inputs before its
first block. This caused the MRS compaction input to drop its very first row,
resulting in missing rows in the output.

This doesn't actually affect real operation because we don't have any
"compacting flush" implemented, but it's a simple to fix now while it's
top-of-mind, in case we implement that later.

Change-Id: Ib7e86557184c138dc54e61c3969af260f7f1c581
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1460
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>