server: add --server_dump_info_path flag to dump bound ports
This allows either the tserver or master to be started with a flag
that causes it to dump its bound port information to a local file.
This will be used as part of an "external minicluster" facility so that
the servers can run on ephemeral ports and prevent conflicts, while
still allowing clients to be configured to connect to them.

The dumped file can either be JSON or raw PB format.

There are no new unit tests included, since the minicluster facility
mentioned above will cover this. Tested manually:

  $ ./build/latest/kudu-master \
    --server_dump_info_path=/tmp/x \
    --master_rpc_bind_addresses=127.0.0.1:0 \
    --master_http_bind_addresses=127.0.0.:0 &
  $ cat /tmp/x
  {
      "bound_rpc_addresses": [
          {
              "host": "127.0.0.1",
              "port": 46168
          }
      ],
      "bound_http_addresses": [
          {
              "host": "0.0.0.0",
              "port": 36372
          }
      ]
  }
  $ ./build/latest/kudu-master \
    --server_dump_info_path=/tmp/x \
    --server_dump_info_format=pb \
    --master_rpc_bind_addresses=127.0.0.1:0 \
    --master_http_bind_addresses=127.0.0.:0 &

  $ ./thirdparty/installed/bin/protoc --decode=kudu.server.ServerStatusPB src/server/server_base.proto  -I src < /tmp/x
  bound_rpc_addresses {
    host: "127.0.0.1"
    port: 50454
  }
  bound_http_addresses {
    host: "0.0.0.0"
    port: 60486
  }

Change-Id: I4e517a56ae9e647e76d90682c208fccadadbfc2a
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1469
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>