KUDU-104. rpc: fix an assertion failure if client messenger shuts down just prior to a new connection
This fixes a race where the client messenger was shutdown after a new outbound
call was put on the reactor thread's queue, but before it had started to assign
a connection. This caused the submission of the negotiation task to the
negotiation threadpool to fail, and thus the Connection object was never put in
the client_conns_ map. Upon exiting the scope, the Connection then got
destructed, which triggered an assertion that connections should not be
destructed without an explicit Shutdown() first.

This patch fixes it by tracking whether the connection is epoll-registered, and
more explicitly verifying in the destructor that it isn't on the epoll loop
anymore.

Change-Id: Ie170c374fb6849b5f52c224c61b83b9de147351e
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1487
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>