KUDU-112, KUDU-103. Fix client-test abort behavior
This changes Batcher::Abort() to delete any buffered operations which
weren't flushed. This prevents an assertion error if you buffer some
operations and never flush, making the behavior consistent with the
case where the operations were still in the midst of DNS resolution.

This is a behavior change for the client, but probably for the better.
See discussion in KUDU-103 for details.

Change-Id: I83188166a3ca2f3f2ad5474f47664aae76aa128e
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1518
Tested-by: jenkins
Reviewed-by: Colin McCabe <cmccabe@cloudera.com>