Add an ExternalMiniCluster implementation
This adds ExternalMiniCluster, a MiniCluster implementation that runs
each daemon in its own process. The purpose of this is to enable fault
testing -- it's easy to send SIGKILL or SIGSTOP/SIGCONT pairs to the
daemons, or otherwise mess with them while they are running.

There is some further work to make this really useful to build tests on
top of, but the basic functionality is verified by a functional test.

TODO in follow-on patches:

* Add functionality like WaitForReplicaCount and WaitForTabletServerCount.
  To implement these, we may need to add some new RPCs to the master to
  expose this information (the internal minicluster relies on white-box
  inspection of the CatalogManager)
* Add some convenience APIs to kill and restart servers, as needed based
  on development of integration tests.

Change-Id: I760bc3111d25a08dbb306e3d11f0ec58bfd1b619
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1533
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>