Undo Deltas: Read/Write path
This patch introduces the full r/w path for undo deltas, including
DiskRowSet r/w integration and compactions/flushes.
As persistent reinserts are not yet possible each time a reinsert
is found it effectively resets all prior history. This can be easily changed
when we add persistent reinserts in the future.
If the last mutation applied to a row is a delete that mutation is kept
as a redo, allowing us never to have to look at undos for the most
current snapshot.

Change-Id: Ieb97286ea5cf565e14f8934e9ed114285fae6bb0
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1551
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>