HybridTime: Encapsulate txid_t's internal representation
txid_t's internal representation is leaked in a lot of places outside
of the Mvcc* classes.
This patch encapsulates the internal representation of txid_t so that
we can drop-in replace another such as HT timestamps.
Mvcc* classes and corresponding test still access the internal
representation directly but these also have to change if another
representation is chosen.

Change-Id: I461f89b5082be76638ec0ed476fac5026c094903
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1573
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>