Add AlterTable to the java client
Adds two new methods to KuduClient, alterTable() to alter a table
by passing a builder like in C++ and syncWaitOnAlterCompletion() which
like its name says is a synchronous method to wait until the alter is done.

Also added was the possiblity to set a default value on columns (was in C++
but not in Java).

Change-Id: Ibad1ddb03fe1152c0ce24dbb405d381f57587f95
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1614
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Tested-by: jenkins