KUDU-58: Make log segment pre-allocation async
This revision implements asynchronous allocation for log segments.

When we reach the maximum segment size, we submit a task to the
executor that creates a file with a temporary name, and begin to
pre-allocate that file to the requested size.

While the new segment is being allocated writes to the existing file
continue beyond the pre-allocated limit (at the cost of requiring an
fsync() each time Sync() is called).

Once the new segment is allocated, it is renamed to the correct
filename, set as the active segment, and initialized with the correct
header.

Performance results (20 writers, YCSB "kudu" workload), additional
logging added on server side to observe the times:

- With async log rolling disabled, the log roll-over delays routinely
exceed 50 ms or get close (~45-49 ms).

- With async log rolling enabled, the delays are now equivalent to the
time it takes to close the previous segment (i.e., time for
ftruncate() + fsync()) and append the header to the new segment --
~30-35 ms on the test machine used.

Further performance gains can be obtained by closing the previous
segment asynchronously after we've already rolled to the new segment:
this would require knowing whether it's safe to perform the
truncate-on-close _after_ we've already started the new segment.

Change-Id: I617c4641022e763b36c33ebee06059efafc83218
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1629
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins