HandleReportedTablet should abort the master if the write on sys-table fail
If we fail the write and we don't abort the master, the in-memory state will
be switched to Created and the tablet will start serving the data.
If at somepoint the master goes down, the Tablet is still marked as "Creating"
in the sys-table, and on restart the tablet will be replaced with a new one
causing data-loss.

Change-Id: I2d3dfcc022f6280275348c12d8ffdb4c3f16abcd
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1630
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins