Make the number of locations returned by GetTableLocations configurable
Right now a Scan would try to fetch all the locations from a table if
it's not configured with start and end key. Or if it has just a start key,
it could try to get the rest of the table even if in reality the goal
was to get some amount of rows without knowing the boundaries.

Adds a slow unit test that has a runtime of about 15 seconds.

Change-Id: I90ddae629028d719b8d93531e7ac38f9f6b4afba
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1633
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Tested-by: jenkins