Undo Deltas: Set Mvcc's initial id != kMin, fix bug in range method and add helper for no-transactions snap
This sets the initial ID generated by Mvcc to be different from kMin, which allows to
create a snapshot, through MvccSnapshot::CreateSnapshotIncludingNoTransactions(),
guaranteed never to look at REDO deltas, even when redos are the first thing that
happens to a tablet (which happens in certain tests).
This patch also fixes a bug in range check useful for undos, which now includes
the boundary, i.e., if a snapshot is MvccSnapshot[committed={T|T < 9} then tx9
still needs to be undone.

Change-Id: Ib4d22b4a75a424bfcbf4db7919721be533d752cf
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1635
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>