HybridTime: Logical clock and Mvcc integration
This patch introduces a logical clock and integrates it with Mvcc manager.
In a single node, single tablet setting this clock generates the same txids
as the MvccManager previously did. This will allow to have the clock be
updated by bootstrap guaranteeing that new transactions after bootstrap
will have new txids, which doesn't happen now if we we bypass tablet server
and definitely won't happen when Log::GC() is working.

This also moves txid_t to the new clock.h header since txid_t is basically
a timestamp.

Change-Id: I919926be4cf1734647ca13968c0dfb15ee0cfb61
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1645
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>