Basic multi-tablet support for the Java client
All the TS operations now support tables with multiple tablets, but not
multiple replicas. Support for tablets not found is implemented for Write
but not properly for Scan. Another problem is that row key comparisons are
not taking into account signed types.

The pom.xml changes were to accommodate the new unit test, they currently
can't run using the same JVM so reuseForks had to be disabled.

ColumnRangePredicatePB got a wrapper class because we need to be able to
extract the row keys from the predicates.

Table pre-splitting was added via the CreateTableBuilder class which
currently only takes string keys.

The key encoder was refactored out of Operation into its own class for
easier reusing.

It doesn't support:
 - using all the replicas (it picks the first one, always)
 - multiple masters

Change-Id: If8386f75c810ae70fe27d3dfaaa0d225c48aa31e
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1665
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>