Make CreateTableOptions::WithSplitKeys() take a const ref
Right now this method is not taking a const ref which prevents
us from from using boost::assign::list_of() which comes handy in tests.
Moreover we shouldn't probably be allowed to mutate the keys vector after
we've passed them to CreateTableOptions (or should use a pointer if we must).

Change-Id: Ia9356a662d3a7c4398a892fc2ad31f37f667acc8
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1672
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>