master: add a ListTabletServers RPC and use it in ExternalMiniCluster
This new administrative RPC lists all of the tablet servers registered on the
master. This is equivalent information to what is shown on the web UI.

This should be useful in management tooling, but it's also useful for
tests that use the ExternalMiniCluster -- they will typically need to
wait until the tablet servers register before performing various actions.

This changes the ExternalMiniCluster Start() method to use the new RPC and
wait until all tablet servers have registered with the master before reporting
the cluster as started.

Change-Id: I50255b6d1960df3260914567658ba81261ebf0da
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1766
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins