HybridTime: HybridClock and related tests
This introduces the HybridClock, another implementation of Clock but
doesn't yet include HT/TT 'external consistency' modes, those are left
for a later patch. As the HybridClock can be used as a regular clock
this changes the clock implementation in TabletServerTest to use it
instead of the LogicalClock (through a flag so that we can also do
it from the command line).

Change-Id: I785e9ff5a9d780c55b0d482044d90fc135ff7614
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1708
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins