HybridTime: Snapshot scans and HT timestamp propagation for scans
This introduces scans in the past (snapshot scans) to tablet server.
Normal snapshot scans are only accepted when the snapshot timestamp is lower
than or equal to clock->Now() as only then is the server sure that no more
writes can happen with a timestamp lower that the snapshot timestamp.

If the client provides no snapshot timestamp the server selects one based
on the current clock time (in which case the above condition always holds).

The difference between snapshot scans and "regular" scans (the default)
is that they are always repeatable, i.e. consecutive snapshot scans with
the same timestamp will always yield the same rows. This is implemented
at the cost of having to wait for in-flight transactions lower than
the snapshot's timestamp to complete.

Additionally the server supports HybridTime timestamp propagation
which allows clients to update a server's clock with a timestamp obtained
from another server (when using external consistency mode CLIENT_PROPAGATED).

Change-Id: If1f074198645978ec6c7f16afe2fef2a9fa84f20
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1718
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>