KUDU-148, KUDU-149. Fix SEGV trying to access a tablet which is bootstrapping
This fixes tablet_service to check the tablet peer state after
looking it up. If the tablet is not running, the tserver responds
with a new TABLET_NOT_RUNNING error code and a ServiceUnavailable
status.

No new tests are included in this patch, but a later patch in this
series does exercise this code path.

Change-Id: Ie745680483b8c65a3ee7e6ffbe9dc931f63e80e9
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1775
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>