KUDU-87. Add a simple linked-list based data integrity system test
This adds a test case similar to HBase's TestLoadAndVerify.
It starts an external minicluster, loads a linked list structure
into a table, restarts the cluster, and re-verifies the structure's
integrity.

There are a couple TODOs in this test case for extensions to the
test that would currently fail. In particular, if we were to crash
the TS while it was in the progress of replaying edits, we'd lose
data and fail verification (part of KUDU-135).

Change-Id: I0ee6a402243aee0f7bf24e0b8e5c0ed50cd7325e
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1776
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>