KUDU-139 - Get rid of missed deltas in the log
This gets rid of missed deltas in the log by making sure that they
are flushed before we update the TabletMetadata with the new rowsets.
Basically this forces a delta flush at the end of ReupdateMissedDeltas()
making sure missed delta data is indeed flushed before the TabletMetadata
is persisted.

Also updates compaction.txt to make it a bit more accurate with the
way that compactions eventually got implemented.

Additional-Author: Todd Lipcon <todd@cloudera.com>
Change-Id: I6f709dd667882939b68c11f089f0b90b0f987823
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1793
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>