KUDU-155 - Increase the waited time tolerance of HybridClockTest.TestWaitUntilAfter_TestCase1
In this test we try to make sure that the clock doen't wait on WaitUntilAfter() as the
Timestamp passed as argument is in the past. Turns out that a 1000 microsecond tolerance
is too low and was making this test fail more often than expected.

This removes the waited time assertion and relies only on the fact that when we wait
we always wait at least 'max_error' (usually two times that), which should be enough.
In the future we should have a mock clock (KUDU-156) to be able to cause this situations
deterministically.

Change-Id: If908994352ab43b799872aa41deb96ecf01bfe73
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1797
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>