KUDU-120 - Pt1 - Add MvccManager::WaitUntilAllCommitted(MvccSnapshot)
This adds WaitUntilAllCommitted() to MvccManager that blocks until all
the in-flight transactions in the snapshot are committed.

This is a step towards KUDU-120 as if we reserve a place on the log
queue before calling commit() on the mvcc transactions and make sure
that all queued commit messages are fsync()ed before proceeding with
the flush, we no longer can have the interleaving described in KUDU-120
and it then becomes safe to reply to the client without fsync()ing the
commit msgs.

This also adds MvccManager::TakeSnapshotAtTimestamp(snap, timestamp) which
allows to get a snapshot that only considers as in-flight transactions those
whose timestamp is lower than 'timestamp'. This is helpful for snapshot
reads since we need to make sure that all transactions whose timestamp
is lower than the snapshot's timestamp are done, i.e. that any two reads
at a snapshot will yield the same rows.

Change-Id: If662f127e70ba696e1cd9e6092efe207182c5531
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1828
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>