KUDU-120 - Pt2 - Allow to enqueue FLUSH_MARKER entries in the log
This adds the ability to enqueue FLUSH_MARKER entries in the log queue that
will not be serialized. The purpose is to allow, through the common log
API, to assert that all messages up to a certain point were in fact
flushed out of the log queue.

The more general goal is to enable us to solve KUDU-120 (only one fsync per
write) by making flushes/compactions first wait for the in-flights to commit,
then wait for all commits messages to be persisted and only then update
the metadata.

Change-Id: I4ca90dd687f02e51735a8223d0e70a2c426c9b37
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1830
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins