Make TestAlterSchema in TabletServerTest insert the rows remotely
Since the test is replaying the log after the alter schema request
and verifying that the rows are in fact there it doesn't make much
sense to skip it to insert the rows by inserting the rows locally.

This was what made me find the bug where we weren't generating new
timestamps on replay, but now that is fixed.

Change-Id: Iaba12943d437314d7281fc4d36193dc0eb596305
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1862
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>