KUDU-128. Fix CHECK failure if RPC server shuts down while connections are incoming
This fixes a race where a connection may be accepted by the AcceptorPool and
then passed to the messenger while it's in the process of shutting down.
Previously we were calling conn->Shutdown() in that case, which fired a CHECK
because it was called from a non-reactor thread.

This Shutdown() call was unnecessary at that point in the lifecycle, since the
connection wasn't registered with epoll and has no calls on it, so has nothing
worth shutting down.

In the process of adding a regression test, I also added a metric for the
number of RPC connections accepted.

Change-Id: I32d7f4444b866128a7c0fb2cf40e2edc6c80cea8
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1863
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>