Make SysTable::SyncWrite return the error on the completion callback
Was getting a swallowed Prepare() error because SysTables::SyncWrite only
returns the error on the WriteResponsePB while that is only set by the usual
callback in tablet server and not by the callback used there.

This makes the method return the error correctly by making the
LatchTransactionCompletionCallback a template that accepts any
response PB that has an 'error' field and setting the error on
the response before the latch count down.

Change-Id: I02f329285ce6b6d8c038810fb4fee59d8173c557
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1865
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Tested-by: jenkins