Fix warning for clipped rows on compaction
This reduces the clipped rows warning to just LOG(WARNING) for the first
row. At the end of the compaction we get a total clipped row count. With
verbosity set to level 2 we get the full detail for every row.

This also moves a very verbose statement in tablet-server-test to VLOG(2).
This makes the log size decrease by a huge amount and reduces the test IO
by several tens of MB's helping to make the test run faster.

Change-Id: Ic2d6063804855d4ff6c687024a53e7a18bae42e6
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1868
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins