[java client] Use IsCreateTable done to verify a table's state
Currently the java client can detect that a table is being created
when it receives no tablets GetTableLocations. This patch is about
handling the case where some tablets were created but not all of them.

While writing this patch, a corner case was uncovered in KuduSession.
When using MANUAL_FLUSH and a freshly created table, the edits that
get applied won't immediatly be in the buffer since we don't know which
tablets they belong to. It means that calling flush() right after that
may not flush any data. This is why TestScannerMultiTablet was modified.

Change-Id: Ia7ccba37fabf8099b5bbee4961210ec2d7acc2e8
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1871
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Tested-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>